feat: implement chunk system
Full implementation of super-perfomant chunk system, which dynamically
loads unneeded objects off from the map.

The implementation was succesfully tested on a map with almost seven
thousand (!) objects with little lags. Thus, it at least runs
extremely smooth with 1-3 thousand objects.

It's planned to make the system more flexible in the future. So far,
this implementation is enough.

Closes #7.